task 1 completed

+++ app/graphql/mutations/post_upvote.rb
@@ -1,9 +1,9 @@
 module Mutations
     class PostUpvote < Mutations::BaseMutation
 
-        argument :post_id, Integer, required: true
+        argument :post_id, Int, required: true
 
-        field :post, Types::PostType, null: false
+        field :post, Types::PostType, null: true
         field :errors, [String], null: false
     
         #field :postId, Integer, null: false
@@ -12,21 +12,28 @@ module Mutations
             require_current_user!
             user = context[:current_user]
           
-            this_post = Post.find(post_id)
+            this_post = Post.find_by(id: post_id)
 
-            upvote = this_post.votes.create(user: user);
+            if this_post.present?
+                 upvote = this_post.votes.create(user: user);
           
-            if upvote.save
-                {
-                    post: this_post,
-                    errors: []
-                }
+                if upvote.save
+                    {
+                        post: this_post,
+                        errors: []
+                    }
+                else
+                    upvote = this_post.votes.find_by(user_id: user.id)
+                    upvote.destroy
+                    {
+                        post: this_post,
+                        errors: user.errors.full_messages
+                    }
+                end
             else
-                upvote = this_post.votes.find_by(user_id: user.id)
-                upvote.destroy
                 {
-                    post: this_post,
-                    errors: user.errors.full_messages
+                    post: nil,
+                    errors: ["record not found"]
                 }
             end
         end
